Refactor: extract dashboard logic into custom hook and lift state
