+++ frontend/src/app/admin/page.tsx
@@ -7,6 +7,8 @@ import StatsCards from './_components/StatsCards/StatsCards';
 import TrendCharts from './_components/TrendCharts/TrendCharts';
 
 
+import { useAdminDashboard } from './useAdminDashboard';
+
 /**
  * 어드민 대시보드 리뉴얼 (기존 스타일 유지 버전)
  * - 좌측 사이드바 레이아웃 유지
@@ -14,6 +16,8 @@ import TrendCharts from './_components/TrendCharts/TrendCharts';
  * - 하단 테이블(ActivityTable)을 트렌드 차트(TrendCharts)로 교체
  */
 export default function AdminDashboardPage() {
+  const { state } = useAdminDashboard();
+
   return (
     <div className="container-sidebar">
       {/* 1. 사이드바 복구 */}
@@ -27,7 +31,7 @@ export default function AdminDashboardPage() {
 
         {/* 요약 카드 섹션 (기존 스타일) */}
         <div style={{ marginTop: '24px' }}>
-          <StatsCards />
+          <StatsCards data={state.data} isLoading={state.isLoading} />
         </div>
 
         {/* 트렌드 분석 섹션 (테이블 대신 차트 배치) */}
@@ -36,7 +40,7 @@ export default function AdminDashboardPage() {
             <h3>서비스 성장 추이</h3>
             <p>최근 7일간의 신규 유저 및 이벤트 데이터 분석</p>
           </div>
-          <TrendCharts />
+          <TrendCharts data={state.data} isLoading={state.isLoading} />
         </div>
         
       </section>
